keep only v2c
