+++ src/smsfusion/__init__.py
@@ -3,14 +3,11 @@ from ._coning_sculling import ConingScullingAlg
 from ._ins import AHRS, VRU, AidedINS, FixedNED, StrapdownINS, gravity
 from ._smoothing import FixedIntervalSmoother
 from ._transforms import quaternion_from_euler
-from ._v2 import AHRSv2a, AHRSv2b, AHRSv2c, AHRSv2d
+from ._v2 import AHRSv2
 
 __all__ = [
     "AHRS",
-    "AHRSv2a",
-    "AHRSv2b",
-    "AHRSv2c",
-    "AHRSv2d",
+    "AHRSv2",
     "AidedINS",
     "benchmark",
     "constants",
