--- src/smsfusion/_v2/__init__.py
@@ -1,6 +1,0 @@
-from ._v2a import AHRSv2a
-from ._v2b import AHRSv2b
-from ._v2c import AHRSv2c
-from ._v2d import AHRSv2d
-
-__all__ = ["AHRSv2a", "AHRSv2b", "AHRSv2c", "AHRSv2d"]
